Added csv reading functionality, Removed unnecessary classes, Changed data fields
Removed TranscriptHandler class
Removed FileReader class
Added StudentRecordSystem Class

Created functionality to read a series of csv files that represent the database that the record system uses.

Read in file contents to objects and stored on the StudentRecordSystem class

Altered many of the variable names for clarity and added data fields in some of the classes to get the file reading functionality started.